refactor(parser): inline ParsableTag derive macro
the derive macro in a local macro caused too many issues (hard to debug,
slow rust-analyzer response times, misleading compile errors, ...)
