+++ derive/src/lib.rs
@@ -1,10 +1,48 @@
 use quote::{quote, quote_spanned};
-use syn::{
-    parse_macro_input, punctuated::Punctuated, AngleBracketedGenericArguments, DeriveInput, Expr,
-    Field, GenericArgument, Ident, Path, PathArguments, PathSegment, Type, TypePath,
-};
+use syn::{parse_macro_input, DeriveInput, Path, PathSegment, Type, TypePath};
 
 extern crate proc_macro;
+
+#[proc_macro_derive(Tag)]
+pub fn tag(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    let name = input.ident;
+    let expanded = match input.data {
+        syn::Data::Enum(syn::DataEnum { variants, .. }) => {
+            let mut start = Vec::new();
+            let mut end = Vec::new();
+            for variant in variants {
+                if variant.fields.len() != 1 {
+                    panic!("Tag is only supported for enum values with exactly one field");
+                }
+                let variant_name = &variant.ident;
+                start.push(quote_spanned! {name.span()=>
+                    #name::#variant_name(inner) => inner.start()
+                });
+                end.push(quote_spanned! {name.span()=>
+                    #name::#variant_name(inner) => inner.end()
+                });
+            }
+            quote! {
+                impl Tag for #name {
+                    fn start(&self) -> Position {
+                        return match self {
+                            #(#start,)*
+                        };
+                    }
+
+                    fn end(&self) -> Position {
+                        return match self {
+                            #(#start,)*
+                        };
+                    }
+                }
+            }
+        }
+        _ => panic!("derive(Tag) is only supported for enums"),
+    };
+    return proc_macro::TokenStream::from(expanded);
+}
 
 #[proc_macro_derive(DocumentNode)]
 pub fn document_node(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -71,293 +109,11 @@ pub fn document_node(input: proc_macro::TokenStream) -> proc_macro::TokenStream 
     return proc_macro::TokenStream::from(expanded);
 }
 
-#[proc_macro_derive(ParsableTag, attributes(tag_definition))]
+#[proc_macro_derive(ParsableTag)]
 pub fn parsable_tag(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
     let name = input.ident;
     let expanded = match input.data {
-        syn::Data::Struct(syn::DataStruct {
-            fields: syn::Fields::Named(syn::FieldsNamed { named, .. }),
-            ..
-        }) => {
-            let mut params: Option<Expr> = None;
-            for attr in &input.attrs {
-                if attr.path().is_ident("tag_definition") {
-                    params = Some(attr.parse_args().unwrap());
-                }
-            }
-            let definition = params.unwrap();
-            let plain_attribute_fields: Vec<&syn::Ident> =
-                attribute_fields_with_option_value(&named, "PlainAttribute");
-            let spel_attribute_fields: Vec<&syn::Ident> =
-                attribute_fields_with_option_of_parsed_attribute_value(&named, "SpelAttribute");
-            quote! {
-                impl #name {
-                    fn try_parse(
-                        parser: &mut TreeParser,
-                        depth_counter: &mut DepthCounter,
-                    ) -> Result<ParsedTag<Self>> {
-                        let parent_node = parser.cursor.node();
-                        let mut errors = Vec::new();
-                        let mut movement = NodeMovement::FirstChild;
-                        depth_counter.bump();
-                        let open_location;
-                        loop {
-                            open_location = match parser.goto(&movement) {
-                                NodeMovingResult::NonExistent | NodeMovingResult::Missing(_) => {
-                                    return Err(anyhow::anyhow!("tag is empty"));
-                                },
-                                NodeMovingResult::Erroneous(node) => {
-                                    return Ok(ParsedTag::Unparsable(
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node),
-                                    ));
-                                },
-                                NodeMovingResult::Superfluous(node) => {
-                                    errors.push(TagError::Superfluous(
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node),
-                                    ));
-                                    movement = NodeMovement::NextSibling;
-                                    continue;
-                                },
-                                NodeMovingResult::Ok(node) => match node_location(node) {
-                                    Location::SingleLine(location) => location,
-                                    location => return Ok(ParsedTag::Unparsable(
-                                        format!(
-                                            "\"{}\" should be on a single line",
-                                            parser.node_text(&node)?,
-                                        ),
-                                        location,
-                                    ))
-                                },
-                            };
-                            break;
-                        }
-                        #(let mut #plain_attribute_fields = None;)*
-                        #(let mut #spel_attribute_fields = None;)*
-                        let mut body = None;
-                        let close_location;
-                        loop {
-                            close_location = match parser.goto(&NodeMovement::NextSibling) {
-                                NodeMovingResult::NonExistent => return Ok(ParsedTag::Unparsable(
-                                    format!("\"{}\" tag is unclosed", #definition.name),
-                                    node_location(parent_node),
-                                )),
-                                NodeMovingResult::Missing(node) if node.kind() == ">" => {
-                                    body = Some(match parser.parse_tag_body()? {
-                                        Some(body) => body,
-                                        None => return Ok(ParsedTag::Unparsable(
-                                            format!("\"{}\" tag is unclosed", #definition.name),
-                                            node_location(node),
-                                        )),
-                                    });
-                                    match #name::parse_closing_tag(parser, &mut errors, node)? {
-                                        Ok(location) => location,
-                                        Err((text, location)) => return Ok(
-                                            ParsedTag::Unparsable(text, location),
-                                        ),
-                                    }
-                                },
-                                NodeMovingResult::Missing(node) if node.kind() == "self_closing_tag_end" => {
-                                    // tree-sitter puts missing "/>" nodes always at the first
-                                    // possible location. in order for completion to work we
-                                    // instead want it to include all following whitespace, so we
-                                    // search for the next node and place it in front of it. if
-                                    // this is the last node we have to manually split the
-                                    // documents text to find "trailing" whitespace, which is not
-                                    // included in any node.
-                                    // however, the error reported must still be on the first
-                                    // possible location such that the quick-fix action inserts it
-                                    // there.
-                                    errors.push(TagError::Missing("/>".to_string(), node_location(node)));
-                                    parser.move_missing_node_past_whitespaces(node)?
-                                },
-                                NodeMovingResult::Missing(node) => {
-                                    return Ok(ParsedTag::Unparsable(
-                                        format!(
-                                            "\"{}\" is missing in \"{}\" tag",
-                                            node.kind(),
-                                            #definition.name
-                                        ),
-                                        node_location(parent_node),
-                                    ));
-                                },
-                                NodeMovingResult::Erroneous(node) => {
-                                    return Ok(ParsedTag::Unparsable(
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node)
-                                    ));
-                                },
-                                NodeMovingResult::Superfluous(node) => {
-                                    errors.push(TagError::Superfluous(
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node),
-                                    ));
-                                    continue;
-                                },
-                                NodeMovingResult::Ok(node) => match node.kind() {
-                                    #(stringify!(#plain_attribute_fields) => {
-                                        #plain_attribute_fields = parser.parse_plain_attribute(),
-                                        continue;
-                                    },)*
-                                    #(stringify!(#spel_attribute_fields) => {
-                                        #spel_attribute_fields = Some(
-                                            stringify!(#spel_attribute_fields)
-                                                .strip_suffix("_attribute")
-                                                .and_then(|n| #definition.attributes.get_by_name(n))
-                                                .map(|d| parser.parse_spel_attribute(&d.r#type))
-                                                .unwrap()?,
-                                        );
-                                        continue;
-                                    },)*
-                                    "self_closing_tag_end" => node_location(node),
-                                    ">" => {
-                                        body = Some(match parser.parse_tag_body()? {
-                                            Some(body) => body,
-                                            None => return Ok(ParsedTag::Unparsable(
-                                                format!("\"{}\" tag is unclosed", #definition.name),
-                                                node_location(node),
-                                            )),
-                                        });
-                                        match #name::parse_closing_tag(parser, &mut errors, node)? {
-                                            Ok(location) => location,
-                                            Err((text, location)) => return Ok(
-                                                ParsedTag::Unparsable(text, location),
-                                            ),
-                                        }
-                                    },
-                                    _ => continue,
-                                },
-                            };
-                            break;
-                        }
-                        let close_location = match close_location {
-                            Location::SingleLine(location) => location,
-                            location => return Ok(ParsedTag::Unparsable(
-                                format!(
-                                    "\"{}\" should be on a single line",
-                                    parser.node_text(&parser.cursor.node())?,
-                                ),
-                                location,
-                            ))
-                        };
-                        let tag = Self {
-                            open_location,
-                            #(#plain_attribute_fields,)*
-                            #(#spel_attribute_fields,)*
-                            body,
-                            close_location,
-                        };
-                        return Ok(match errors.is_empty() {
-                            true => ParsedTag::Valid(tag),
-                            false => ParsedTag::Erroneous(tag, errors),
-                        });
-                    }
-
-                    fn parse_closing_tag(
-                        parser: &mut TreeParser,
-                        errors: &mut Vec<TagError>,
-                        body_open_node: tree_sitter::Node<'_>,
-                    ) -> Result<Result<Location, (String, Location)>> {
-                        loop {
-                            return Ok(Ok(match parser.goto(&NodeMovement::Current) {
-                                NodeMovingResult::Missing(node) => {
-                                    // tree-sitter puts missing "<{tag}/>" nodes always after all
-                                    // its siblings, which is ideal for completion.
-                                    // however, the error reported must be on the first possible
-                                    // location such that the quick-fix action inserts it there.
-                                    let start_position = body_open_node.start_position();
-                                    errors.push(TagError::Missing(
-                                        format!("</{}>", #definition.name),
-                                        Location::SingleLine(SingleLineLocation {
-                                            char: start_position.column + 1,
-                                            line: start_position.row,
-                                            length: 0,
-                                        }),
-                                    ));
-                                    parser.move_missing_node_past_whitespaces(node)?
-                                },
-                                NodeMovingResult::Erroneous(node) => {
-                                    return Ok(Err((
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node),
-                                    )));
-                                },
-                                NodeMovingResult::Superfluous(node) => {
-                                    errors.push(TagError::Superfluous(
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node),
-                                    ));
-                                    continue;
-                                },
-                                NodeMovingResult::Ok(node) if node.kind().ends_with("_tag_close") => {
-                                    node_location(node)
-                                },
-                                NodeMovingResult::Ok(node) => {
-                                    return Ok(Err((
-                                        parser.node_text(&node)?.to_string(),
-                                        node_location(node),
-                                    )));
-                                },
-                                // cannot happen
-                                NodeMovingResult::NonExistent => continue,
-                            }));
-                        }
-                    }
-                }
-
-                impl ParsableTag for #name {
-                    fn parse(parser: &mut TreeParser) -> Result<ParsedTag<Self>> {
-                        let mut depth_counter = DepthCounter::new();
-                        let result = #name::try_parse(parser, &mut depth_counter);
-                        for _ in 0..depth_counter.get() {
-                            parser.cursor.goto_parent();
-                        }
-                        return result;
-                    }
-
-                    fn definition(&self) -> TagDefinition {
-                        return #definition;
-                    }
-
-                    fn open_location(&self) -> &SingleLineLocation {
-                        return &self.open_location;
-                    }
-
-                    fn close_location(&self) -> &SingleLineLocation {
-                        return &self.close_location;
-                    }
-
-                    fn body(&self) -> &Option<TagBody> {
-                        return &self.body;
-                    }
-
-                    fn spel_attributes(&self) -> Vec<(&str, &ParsedAttribute<SpelAttribute>)> {
-                        let mut attributes = Vec::new();
-                        #(
-                            match self.#spel_attribute_fields.as_ref() {
-                                Some(field) => attributes.push(
-                                    (stringify!(#spel_attribute_fields), field)
-                                ),
-                                None => (),
-                            };
-                        )*
-                        return attributes;
-                    }
-
-                    fn spel_attribute(&self, name: &str) -> Option<&ParsedAttribute<SpelAttribute>> {
-                        return match format!("{}_attribute", name).as_str() {
-                            #(
-                                stringify!(#spel_attribute_fields) =>
-                                    self.#spel_attribute_fields.as_ref(),
-                            )*
-                            _ => None,
-                        };
-                    }
-                }
-            }
-        }
         syn::Data::Enum(syn::DataEnum { variants, .. }) => {
             let mut definition = Vec::new();
             let mut open_location = Vec::new();
@@ -436,73 +192,6 @@ pub fn parsable_tag(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
         _ => panic!("derive(ParsableTag) is only supported for structs and enums"),
     };
     return proc_macro::TokenStream::from(expanded);
-}
-
-fn attribute_fields_with_option_value<'a, T>(
-    fields: &'a Punctuated<Field, T>,
-    value_type_name: &str,
-) -> Vec<&'a Ident> {
-    fields
-        .iter()
-        .filter(|field| is_type_option_of(field, &value_type_name))
-        .filter_map(|field| is_attribute_field(field))
-        .collect()
-}
-
-fn attribute_fields_with_option_of_parsed_attribute_value<'a, T>(
-    fields: &'a Punctuated<Field, T>,
-    value_type_name: &str,
-) -> Vec<&'a Ident> {
-    fields
-        .iter()
-        .filter(|field| is_type_option_of_parsed_attribute_of(field, &value_type_name))
-        .filter_map(|field| is_attribute_field(field))
-        .collect()
-}
-
-fn is_attribute_field(field: &Field) -> Option<&Ident> {
-    field.ident.as_ref().filter(|ident| {
-        ident
-            .span()
-            .source_text()
-            .is_some_and(|name| name.ends_with("_attribute"))
-    })
-}
-
-fn is_type_option_of(field: &Field, type_name: &str) -> bool {
-    return get_first_generic_type_of(&field.ty, "Option")
-        .and_then(|a| match a {
-            Type::Path(TypePath { path, .. }) => Some(path),
-            _ => None,
-        })
-        .is_some_and(|path| path.is_ident(type_name));
-}
-
-fn is_type_option_of_parsed_attribute_of(field: &Field, type_name: &str) -> bool {
-    return get_first_generic_type_of(&field.ty, "Option")
-        .and_then(|r#type| get_first_generic_type_of(r#type, "ParsedAttribute"))
-        .and_then(|a| match a {
-            Type::Path(TypePath { path, .. }) => Some(path),
-            _ => None,
-        })
-        .is_some_and(|path| path.is_ident(type_name));
-}
-
-fn get_first_generic_type_of<'a>(r#type: &'a Type, type_name: &str) -> Option<&'a Type> {
-    return get_type(r#type)
-        .filter(|r#type| get_type_name(r#type).is_some_and(|name| name == type_name))
-        .and_then(get_first_generic_type);
-}
-
-fn get_first_generic_type(path: &PathSegment) -> Option<&Type> {
-    if let PathArguments::AngleBracketed(AngleBracketedGenericArguments { args, .. }) =
-        &path.arguments
-    {
-        if let Some(GenericArgument::Type(r#type)) = args.first() {
-            return Some(r#type);
-        }
-    }
-    return None;
 }
 
 fn get_type(r#type: &Type) -> Option<&PathSegment> {
